requirements.txt
openpyxl~=3.1.2
pandas~=2.0.3
fastapi~=0.101.1
python-dotenv~=1.0.0
starlette~=0.27.0
pydantic~=2.2.1
uvicorn~=0.23.2
azure.storage.blob~=12.17.0
numpy~=1.25.2